refactor: 카드 유효성 검사 에러 상태를 custom 훅으로 분리
