+++ src/components/cardInfoForm/CardForm/CardForm.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useReducer, useState } from 'react';
+import React, { useState } from 'react';
 import CardInputSection from '../CardInputSection/CardInputSection';
 import CardNumberField from '../CardNumberField/CardNumberField';
 import CardCVCField from '../CardCVCField/CardCVCField';
@@ -6,163 +6,8 @@ import { CARD_IFNO_INPUT_STEP } from '../../../App';
 import CardCompanySelect from '../CardCompanySelect/CardCompanySelect';
 import { CARD_COMPANY_NAME } from '../../constants/cardCompany';
 import CardValidityPeriodField from '../CardValidityPeriodField/CardValidityPeriodField';
-import { checkValideDate } from '../../../utils/checkValideDate';
 import CardPasswordField from '../CardPasswordField/CardPasswordField';
-
-const VALIDITY_PERIOD = {
-  MAX_LENGTH: 2,
-};
-
-interface Action {
-  type:
-    | 'VALIDATE_CARD_NUMBER'
-    | 'VALIDATE_CARD_CVC'
-    | 'VALIDATE_CARD_VALIDITY_PERIOD'
-    | 'VALIDATE_CARD_PASSWORD';
-  payload: {
-    cardNumber?: string[];
-    cardCVC?: string;
-    cardValidityPeriod?: {
-      month: string;
-      year: string;
-      type: 'month' | 'year';
-    };
-    cardPassword?: string;
-  };
-}
-
-interface FormState {
-  isCardNumberError: boolean[];
-  isErrorCardValidityPeriod: {
-    month: boolean;
-    year: boolean;
-  };
-  isCardCVCError: boolean;
-  isErrorCardPassword: boolean;
-  cardNumberErrorMessage: string;
-  cardCVCErrorMessage: string;
-  cardValidityMessage: string;
-  cardPasswordErrorMessage: string;
-}
-
-const initialState: FormState = {
-  isCardNumberError: [false, false, false, false],
-  isErrorCardValidityPeriod: { month: false, year: false },
-  isCardCVCError: false,
-  isErrorCardPassword: false,
-  cardNumberErrorMessage: '',
-  cardCVCErrorMessage: '',
-  cardValidityMessage: '',
-  cardPasswordErrorMessage: '',
-};
-
-type ExpiryField = 'month' | 'year';
-
-interface ValidationResult {
-  message: string;
-  field: ExpiryField | null;
-}
-
-const validateExpiry = (
-  monthStr: string,
-  yearStr: string,
-  focueType: ExpiryField,
-): ValidationResult => {
-  const INVALID_DATE_MSG = '현재보다 이전값을 유효기간으로 선택할 수 없습니다.';
-  const FORMAT_MSG = 'MM형식으로 입력해주세요. (ex. 01)';
-  const MONTH_RANGE_MSG = '1~12사이의 올바른 월을 입력해 주세요.';
-
-  const month = Number(monthStr);
-  const year = Number(yearStr);
-  const currentYear = Number(new Date().getFullYear().toString().slice(2));
-
-  if (focueType === 'month') {
-    if (monthStr.length < 2) return { message: FORMAT_MSG, field: 'month' };
-    if (month < 1 || month > 12)
-      return { message: MONTH_RANGE_MSG, field: 'month' };
-  }
-  if (focueType === 'year') {
-    if (yearStr.length < 2) return { message: FORMAT_MSG, field: 'year' };
-    if (year < currentYear) return { message: INVALID_DATE_MSG, field: 'year' };
-  }
-
-  const invalidField = checkValideDate(monthStr, yearStr);
-  if (invalidField) {
-    return { message: INVALID_DATE_MSG, field: invalidField };
-  }
-
-  if (focueType === 'month') {
-    if (year < currentYear) return { message: INVALID_DATE_MSG, field: 'year' };
-    if (yearStr.length < 2) return { message: FORMAT_MSG, field: 'year' };
-  }
-
-  if (focueType === 'year') {
-    if (month < 1 || month > 12)
-      return { message: MONTH_RANGE_MSG, field: 'month' };
-    if (monthStr.length < 2) return { message: FORMAT_MSG, field: 'month' };
-  }
-
-  return { message: '', field: null };
-};
-
-function formReducer(state: FormState, action: Action): FormState {
-  switch (action.type) {
-    case 'VALIDATE_CARD_NUMBER': {
-      const nums = action.payload.cardNumber || [];
-      const errors = nums.map((v) => v !== '' && v.length !== 4);
-
-      const message = errors.some(Boolean)
-        ? '카드 번호는 4자리씩 입력해야 합니다.'
-        : '';
-      return {
-        ...state,
-        isCardNumberError: errors,
-        cardNumberErrorMessage: message,
-      };
-    }
-    case 'VALIDATE_CARD_CVC': {
-      const cvc = action.payload.cardCVC || '';
-      const isError = cvc.length !== 3;
-      const msg = isError ? 'CVC는 3자리 숫자여야 합니다.' : '';
-      return {
-        ...state,
-        isCardCVCError: isError,
-        cardCVCErrorMessage: msg,
-      };
-    }
-    case 'VALIDATE_CARD_VALIDITY_PERIOD': {
-      const { type, month, year } = action.payload.cardValidityPeriod!;
-
-      const { message, field } = validateExpiry(month, year, type);
-
-      const isError = field
-        ? {
-            month: type === 'month' ? false : false,
-            year: false,
-            [field]: true,
-          }
-        : { month: false, year: false };
-
-      return {
-        ...state,
-        isErrorCardValidityPeriod: isError,
-        cardValidityMessage: message,
-      };
-    }
-    case 'VALIDATE_CARD_PASSWORD': {
-      const password = action.payload.cardPassword || '';
-      const isError = password.length > 0 && password.length < 2;
-
-      return {
-        ...state,
-        isErrorCardPassword: isError,
-        cardPasswordErrorMessage: isError ? '비밀번호는 두자리 입니다.' : '',
-      };
-    }
-    default:
-      return state;
-  }
-}
+import useCardValidationError from '../../../hooks/useCardValidationError';
 
 interface CardFormProps {
   cardNumber: string[];
@@ -198,8 +43,6 @@ function CardForm({
   handleChangeValidityPeriod,
   handleChangeCardPassword,
 }: CardFormProps) {
-  const [state, dispatch] = useReducer(formReducer, initialState);
-
   const [show, setShow] = useState({
     cardCompany: false,
     cvc: false,
@@ -212,6 +55,19 @@ function CardForm({
     setShow((prev) => ({ ...prev, [step]: true }));
   };
 
+  const {
+    state,
+    validateCardNumber,
+    validateCardCVC,
+    validatePeriod,
+    validatePassword,
+  } = useCardValidationError({
+    cardNumber,
+    cardCVC,
+    cardValidityPeriod,
+    showNextStep,
+  });
+
   const onChangeCardNumber = (
     e: React.ChangeEvent<HTMLInputElement>,
     i: number,
@@ -219,15 +75,12 @@ function CardForm({
     handleChangeCardNumber(e, i);
     const copy = [...cardNumber];
     copy[i] = e.target.value;
-    dispatch({ type: 'VALIDATE_CARD_NUMBER', payload: { cardNumber: copy } });
+    validateCardNumber(copy);
   };
 
   const onChangeCardCVC = (e: React.ChangeEvent<HTMLInputElement>) => {
     handleChangeCVC(e);
-    dispatch({
-      type: 'VALIDATE_CARD_CVC',
-      payload: { cardCVC: e.target.value },
-    });
+    validateCardCVC(e.target.value);
   };
 
   const onChangeCardValidityPeriod = (
@@ -240,51 +93,13 @@ function CardForm({
     if (type === 'month') copy.month = e.target.value;
     if (type === 'year') copy.year = e.target.value;
 
-    dispatch({
-      type: 'VALIDATE_CARD_VALIDITY_PERIOD',
-      payload: {
-        cardValidityPeriod: {
-          ...copy,
-          type,
-        },
-      },
-    });
+    validatePeriod(copy.year, copy.month, type);
   };
 
   const onChangeCardPassword = (e: React.ChangeEvent<HTMLInputElement>) => {
     handleChangeCardPassword(e);
-    dispatch({
-      type: 'VALIDATE_CARD_PASSWORD',
-      payload: { cardPassword: e.target.value },
-    });
+    validatePassword(e.target.value);
   };
-
-  useEffect(() => {
-    if (
-      state.isCardNumberError.every((e) => !e) &&
-      cardNumber.every((e) => e !== '')
-    ) {
-      showNextStep(CARD_IFNO_INPUT_STEP.cardCompany);
-    }
-
-    if (
-      Object.values(state.isErrorCardValidityPeriod).every((e) => !e) &&
-      Object.values(cardValidityPeriod).every((e) => e !== '')
-    ) {
-      showNextStep(CARD_IFNO_INPUT_STEP.cvc);
-    }
-
-    if (cardCVC !== '' && !state.isCardCVCError) {
-      showNextStep(CARD_IFNO_INPUT_STEP.password);
-    }
-  }, [
-    cardNumber,
-    cardValidityPeriod,
-    cardCVC,
-    state.isCardNumberError,
-    state.isErrorCardValidityPeriod,
-    state.isCardCVCError,
-  ]);
 
   return (
     <form>
